% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlist.r
\name{inlist}
\alias{inlist}
\title{Presumably more convinient version of 'apply' for lists with table like structure.}
\usage{
inlist(l, extractor, selector)
}
\arguments{
\item{l}{list}

\item{extractor}{Predicate to filter list elements before applying.}

\item{selector}{}
}
\value{
List
}
\description{
List element is bound to \code{extractor} and \code{selector} enviroments so its named elements available as variables
}
