% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlist.r
\name{inlist}
\alias{inlist}
\title{Select and apply expression to elements of table like lists.}
\usage{
inlist(l, extractor, applicator, fallback)
}
\arguments{
\item{l}{list (or NULL)}

\item{extractor}{Predicate expression to filter list elements before applying evaluated in the environment of list's element}

\item{applicator}{Expression applied to each element of the list evaluated in the element's environment}

\item{fallback}{Optional. If provided use this value as a fall back in case some variables (prefixed with dot) are not fount in the list's element environment. Otherwise (the default) those elements will be ignored and not included to results}
}
\value{
List of filtered with \code{extractor} elements with values returned by \code{applicator}. Unbound expressions are omitted (if \code{fallback} is not provided) so list might be shorter.
}
\description{
Each list element is attached to both \code{extractor} and \code{applicator} environment so its own named elements are available as variables prefixed with "." (dot character).
}
\details{
\itemize{
\item \code{.} - current list's element
\item \code{..} - (filtered) list (in \code{extractor} environment it is the same as initial list)
\item \code{.n} - (filtered) list element's index
\item \code{.N} - (filtered) list's length
\item \code{...} - initial list (\code{l})
\item \code{..n} - initial list element's index
\item \code{..N} - initial list's length
\item \code{._(x, fb)} - function that returns the fallback \code{fb} if some of the variables in \code{x} are unbound (i.e., due to non existing list elements)
}
}
