% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlist.r
\name{inlist}
\alias{inlist}
\title{Presumably more convenient version of 'apply' for lists with table like structure.}
\usage{
inlist(l, extractor, applicator, fallback)
}
\arguments{
\item{l}{list (or NULL)}

\item{extractor}{Predicate expression to filter list elements before applying evaluated in the environment of list's element}

\item{applicator}{Expression applied to each element of the list evaluated in the element's environment}

\item{fallback}{Optional. If provided use this value as a fall back in case some variables (prefixed with dot) are not fount in the list's element environment. Otherwise (the default) those elements will be ignored and not included to results}
}
\value{
List of filtered with \code{extractor} elements with values returned by \code{applicator}. Unbound expressions are omitted (if \code{fallback} is not provided) so list might be shorter.
}
\description{
List element is bound to \code{extractor} and \code{applicator} environments so its named elements available as variables prefixed with dots
}
